fix: put motion cards into separate client file
