+++ src/components/events/hover-card.tsx
@@ -1,19 +1,14 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
+import { AnimatePresence } from "framer-motion";
+import { debounce } from "lodash";
+
+import { Card, CardDescription, CardTitle } from "@/components/ui/card";
 import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
   MotionCard,
   MotionCardContent,
   MotionCardHeader,
-} from "@/components/ui/card";
-import type { EventType } from "@db/types";
-import { debounce } from "lodash";
+} from "@/components/ui/motion-card";
 import { cn } from "@/lib/utils";
-import { AnimatePresence, motion, Variants } from "framer-motion";
 
 type Props = {
   // event: Pick<EventType, "title" | "description" | "color">;
@@ -49,8 +44,8 @@ const HoverCard = ({ className, ...props }: Props) => {
     <MotionCard
       className={cn(
         className,
-        "w-full p-4 select-none overflow-hidden relative rounded-2xl transition-colors cursor-pointer",
-        { "bg-blue-900": expanded },
+        "relative w-full cursor-pointer select-none overflow-hidden rounded-2xl p-4 transition-colors",
+        { "bg-blue-900": expanded }
       )}
       style={{ height: expanded ? "auto" : "fit-content" }}
       onTap={handleTap}
@@ -64,14 +59,12 @@ const HoverCard = ({ className, ...props }: Props) => {
         stiffness: 700,
         damping: 30,
       }}
-      layout
-    >
+      layout>
       <MotionCardHeader
         layout
         className={cn("p-4", {
           "pb-6": expanded,
-        })}
-      >
+        })}>
         <CardTitle>Card Title</CardTitle>
         <CardDescription>Card Description</CardDescription>
       </MotionCardHeader>
